pyproject.toml
[project]
name = "safeclaw"
version = "0.1.0"
description = "A privacy-first personal automation assistant - no GenAI required"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "SafeClaw Contributors"}
]
keywords = ["automation", "assistant", "privacy", "self-hosted", "webhook", "crawler"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]

dependencies = [
    # Core
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Web crawling
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.0.0",

    # Summarization (non-AI)
    "sumy>=0.11.0",
    "nltk>=3.8.0",

    # RSS feeds
    "feedparser>=6.0.0",

    # Text processing
    "dateparser>=1.2.0",
    "rapidfuzz>=3.5.0",

    # Webhooks & API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Notifications
    "desktop-notifier>=3.5.0",

    # Sentiment analysis (rule-based, no ML)
    "vaderSentiment>=3.3.2",

    # Calendar support
    "icalendar>=5.0.0",

    # Document reading
    "pymupdf>=1.23.0",  # PDF reading
    "python-docx>=1.1.0",  # DOCX reading

    # Image handling (lightweight)
    "pillow>=10.0.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=20.7"]
discord = ["discord.py>=2.3.0"]
slack = ["slack-sdk>=3.26.0"]
matrix = ["matrix-nio>=0.24.0"]
email = ["aiosmtplib>=3.0.0", "aioimaplib>=1.0.0"]
smarthome = ["phue>=1.1", "paho-mqtt>=1.6.0"]
browser = ["playwright>=1.41.0"]
caldav = ["caldav>=1.3.0"]  # For CalDAV server sync

# ML features (optional - heavy dependencies)
nlp = ["spacy>=3.7.0", "langdetect>=1.0.9"]  # NER, ~50MB
vision = ["ultralytics>=8.0.0", "pytesseract>=0.3.10"]  # YOLO+OCR, ~2GB (needs PyTorch)
ocr = ["pytesseract>=0.3.10"]  # OCR only, lightweight

all = [
    "safeclaw[telegram,discord,slack,matrix,email,smarthome,browser,caldav]"
]
ml = ["safeclaw[nlp,vision]"]  # All ML features
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
safeclaw = "safeclaw.cli:main_cli"

[project.urls]
Homepage = "https://github.com/safeclaw/safeclaw"
Documentation = "https://github.com/safeclaw/safeclaw#readme"
Repository = "https://github.com/safeclaw/safeclaw"
Issues = "https://github.com/safeclaw/safeclaw/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/safeclaw"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
